+++ apps/web/src/views/profile/TradePage.tsx
@@ -1,31 +1,21 @@
 'use client';
 // 마이페이지 - 판매 내역 페이지
-import { useState } from 'react';
 
 import { AuctionItemList } from '@/components/common/auction-card';
-import { TabListFilter } from '@/components/common/tab';
+import FilterTab from '@/components/common/tab/filter';
 
-import { AUCTION_TABS, AUCTION_TABS_EXTENDED, TAB_STATUS_MAP, TabId } from '@/lib/constants/tabs';
+import { AUCTION_TABS_EXTENDED, TAB_STATUS_MAP } from '@/lib/constants/tabs';
+import { useFilterStore } from '@/lib/zustand/store';
+
+import { TabId } from '@/types/filter';
 
 const TradeListPage = () => {
-  const [selectedTab, setSelectedTab] = useState<TabId>(AUCTION_TABS[0]?.id || 'all');
-
-  const handleTabChange = (tabId: string) => {
-    setSelectedTab(tabId as TabId);
-  };
+  const selectedTab = useFilterStore((store) => store.selectedItems.trade) as TabId;
 
   return (
     <div>
       {/* 상단 헤더 */}
-      {/* <div className="absolute left-0 top-0 z-10 w-full">
-        <Header title="판매 내역" />
-      </div> */}
-
-      <TabListFilter
-        items={AUCTION_TABS_EXTENDED}
-        selectedCategoryId={selectedTab}
-        onCategoryChange={handleTabChange}
-      />
+      <FilterTab filterKey="trade" items={AUCTION_TABS_EXTENDED} />
       <AuctionItemList selectedStatuses={TAB_STATUS_MAP[selectedTab]} includeCompleted={true} />
     </div>
   );
